Correct artifact coordinates and package
