--- client/src/main/kotlin/org/octopusden/octopus/releasemanagement/client/ReleaseManagementServiceClient.kt
@@ -1,9 +1,0 @@
-package org.octopusden.octopus.releasemanagement.client
-
-import feign.RequestLine
-import org.octopusden.octopus.releasemanagement.client.common.dto.ServiceInfoDTO
-
-interface ReleaseManagementServiceClient {
-    @RequestLine("GET /actuator/info")
-    fun getServiceInfo(): ServiceInfoDTO
-}
