--- client/src/main/kotlin/org/octopusden/octopus/releasemanagement/client/impl/ReleaseManagementServiceClientParametersProvider.kt
@@ -1,6 +1,0 @@
-package org.octopusden.octopus.releasemanagement.client.impl
-
-interface ReleaseManagementServiceClientParametersProvider {
-    fun getApiUrl(): String
-    fun getTimeRetryInMillis(): Int
-}
