+++ client/src/main/kotlin/org/octopusden/octopus/releasemanagementservice/client/impl/ReleaseManagementServiceClientParametersProvider.kt
@@ -1,0 +1,6 @@
+package org.octopusden.octopus.releasemanagementservice.client.impl
+
+interface ReleaseManagementServiceClientParametersProvider {
+    fun getApiUrl(): String
+    fun getTimeRetryInMillis(): Int
+}
