--- common/src/main/kotlin/org/octopusden/octopus/releasemanagement/client/common/dto/ErrorResponse.kt
@@ -1,3 +1,0 @@
-package org.octopusden.octopus.releasemanagement.client.common.dto
-
-data class ErrorResponse(val errorCode: ReleaseManagementServiceErrorCode, val errorMessage: String)
