--- common/src/main/kotlin/org/octopusden/octopus/releasemanagement/client/common/dto/ServiceInfoDTO.kt
@@ -1,5 +1,0 @@
-package org.octopusden.octopus.releasemanagement.client.common.dto
-
-data class ServiceInfoDTO(val build: Build) {
-    data class Build(val version: String)
-}
