--- common/src/main/kotlin/org/octopusden/octopus/releasemanagement/client/common/exception/ArgumentsNotCompatibleException.kt
@@ -1,3 +1,0 @@
-package org.octopusden.octopus.releasemanagement.client.common.exception
-
-class ArgumentsNotCompatibleException(message: String): ReleaseManagementServiceException(message)
