--- common/src/main/kotlin/org/octopusden/octopus/releasemanagement/client/common/exception/HandledResponseCode.kt
@@ -1,14 +1,0 @@
-package org.octopusden.octopus.releasemanagement.client.common.exception
-
-import org.octopusden.octopus.releasemanagement.client.common.dto.ErrorResponse
-
-enum class HandledResponseCode(val responseCode: Int, val exceptionFunction: (ErrorResponse) -> Nothing) {
-    BAD_REQUESTS(400, { response -> throw NotFoundException(response.errorMessage) }),
-    INTERNAL_SERVER_ERROR(500, { response -> throw IllegalStateException(response.errorMessage) });
-
-    companion object {
-        fun mappedException(responseCode: Int): HandledResponseCode? {
-            return values().find { it.responseCode == responseCode }
-        }
-    }
-}
