--- common/src/main/kotlin/org/octopusden/octopus/releasemanagement/client/common/exception/NotFoundException.kt
@@ -1,3 +1,0 @@
-package org.octopusden.octopus.releasemanagement.client.common.exception
-
-class NotFoundException(message: String) : ReleaseManagementServiceException(message)
