--- common/src/main/kotlin/org/octopusden/octopus/releasemanagement/client/common/exception/ReleaseManagementServiceException.kt
@@ -1,3 +1,0 @@
-package org.octopusden.octopus.releasemanagement.client.common.exception
-
-abstract class ReleaseManagementServiceException(message: String) : RuntimeException(message)
