+++ common/src/main/kotlin/org/octopusden/octopus/releasemanagementservice/client/common/dto/ErrorResponse.kt
@@ -1,0 +1,3 @@
+package org.octopusden.octopus.releasemanagementservice.client.common.dto
+
+data class ErrorResponse(val errorCode: ReleaseManagementServiceErrorCode, val errorMessage: String)
