+++ common/src/main/kotlin/org/octopusden/octopus/releasemanagementservice/client/common/dto/ServiceInfoDTO.kt
@@ -1,0 +1,5 @@
+package org.octopusden.octopus.releasemanagementservice.client.common.dto
+
+data class ServiceInfoDTO(val build: Build) {
+    data class Build(val version: String)
+}
