+++ common/src/main/kotlin/org/octopusden/octopus/releasemanagementservice/client/common/exception/ArgumentsNotCompatibleException.kt
@@ -1,0 +1,3 @@
+package org.octopusden.octopus.releasemanagementservice.client.common.exception
+
+class ArgumentsNotCompatibleException(message: String): ReleaseManagementServiceException(message)
