+++ common/src/main/kotlin/org/octopusden/octopus/releasemanagementservice/client/common/exception/NotFoundException.kt
@@ -1,0 +1,3 @@
+package org.octopusden.octopus.releasemanagementservice.client.common.exception
+
+class NotFoundException(message: String) : ReleaseManagementServiceException(message)
