+++ common/src/main/kotlin/org/octopusden/octopus/releasemanagementservice/client/common/exception/ReleaseManagementServiceException.kt
@@ -1,0 +1,3 @@
+package org.octopusden.octopus.releasemanagementservice.client.common.exception
+
+abstract class ReleaseManagementServiceException(message: String) : RuntimeException(message)
