--- ft/src/ft/kotlin/org/octopusden/octopus/releasemanagement/ActuatorTest.kt
@@ -1,8 +1,0 @@
-package org.octopusden.octopus.releasemanagement
-
-import org.octopusden.octopus.releasemanagement.client.common.dto.ServiceInfoDTO
-
-class ActuatorTest : BaseReleaseManagementServiceFuncTest() {
-
-    override fun getServiceInfo(): ServiceInfoDTO = client.getServiceInfo()
-}
