+++ ft/src/ft/kotlin/org/octopusden/octopus/releasemanagementservice/ActuatorTest.kt
@@ -1,0 +1,8 @@
+package org.octopusden.octopus.releasemanagementservice
+
+import org.octopusden.octopus.releasemanagementservice.client.common.dto.ServiceInfoDTO
+
+class ActuatorTest : BaseReleaseManagementServiceFuncTest() {
+
+    override fun getServiceInfo(): ServiceInfoDTO = client.getServiceInfo()
+}
