--- server/src/main/kotlin/org/octopusden/octopus/releasemanagement/ReleaseManagementServiceApplication.kt
@@ -1,13 +1,0 @@
-package org.octopusden.octopus.releasemanagement
-
-import org.springframework.boot.SpringApplication
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan
-
-@SpringBootApplication
-@ConfigurationPropertiesScan
-class ReleaseManagementServiceApplication
-
-fun main(args: Array<String>) {
-    SpringApplication.run(ReleaseManagementServiceApplication::class.java, *args)
-}
