--- server/src/main/kotlin/org/octopusden/octopus/releasemanagement/config/WebConfig.kt
@@ -1,12 +1,0 @@
-package org.octopusden.octopus.releasemanagement.config
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.servlet.config.annotation.ViewControllerRegistry
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-
-@Configuration
-class WebConfig : WebMvcConfigurer {
-    override fun addViewControllers(registry: ViewControllerRegistry) {
-        registry.addRedirectViewController("/", "swagger-ui/index.html")
-    }
-}
