+++ server/src/main/kotlin/org/octopusden/octopus/releasemanagementservice/ReleaseManagementServiceApplication.kt
@@ -1,0 +1,13 @@
+package org.octopusden.octopus.releasemanagementservice
+
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan
+
+@SpringBootApplication
+@ConfigurationPropertiesScan
+class ReleaseManagementServiceApplication
+
+fun main(args: Array<String>) {
+    SpringApplication.run(ReleaseManagementServiceApplication::class.java, *args)
+}
