+++ server/src/main/kotlin/org/octopusden/octopus/releasemanagementservice/config/WebConfig.kt
@@ -1,0 +1,12 @@
+package org.octopusden.octopus.releasemanagementservice.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
+
+@Configuration
+class WebConfig : WebMvcConfigurer {
+    override fun addViewControllers(registry: ViewControllerRegistry) {
+        registry.addRedirectViewController("/", "swagger-ui/index.html")
+    }
+}
