+++ server/src/test/resources/application-test.yml
@@ -6,4 +6,4 @@ management:
 
 logging:
   level:
-    org.octopusden.octopus.releasemanagement.controller: trace
+    org.octopusden.octopus.releasemanagementservice.controller: trace
